Create heading cells in table with row spans with more straightforward XSLT code
Instead of counting the number of rows explicitly beforehand, generate the output
in two passes.  First, generate the content rows, then go back and prepend the
head cell once the number of content rows is known.

This technique was also possible in XSLT 1 but XSLT 2 just makes it a lot easier.