Use HTML superscripts instead of Unicode superscripts for Jyutping tone numbers
The former is obviously easier to type.  But the primary reason for this change
is the font display for the Unicode superscripts has a serious problem.  The
automatic subsetting of the font that Google's server does to the Noto font
excludes some of the critical superscript characters, so they are displayed in
another font which looks terrible.  Also we want to have a graceful fallback
in case Noto (as a Web font) is not available; the other fonts that the user
is viewing the document with may also have suboptimal Unicode superscripts.

It happens that the HTML superscripts are a little bigger (at least in Firefox)
than the Unicode superscripts, which is a plus.  It may also be possible for
us to customize the size of the superscripts later in CSS.